Iot-freshness-tracker-master-latest-version/requirements.txt
flask
flask-socketio
psycopg2-binary
eventlet
flask-cors